Comments and fixed ternary operator errors
Also fixed some spelling mistakes in discountdifference in later lines! Ughh! Stupid names!

But the ternary operator was originally only for the innerHTML not for the actual calculation so it was always subtracting 10%! Hard to catch as it would show no discount, but still take it off of the total! Whoops, now the innerHTML is showing the variable and the ternary is moved up and works on the math! Making the varibale always accurate!